change architecture - files and folders
